+++ src/apis/channel/getChannelInform.tsx
@@ -1,6 +1,6 @@
 import axiosInstance from '../api'
 
-import { GET_CHANNEL_LIST_PATH } from '@/utils/api_paths'
+import { GET_CHANNEL_LIST_PATH } from '@/apis/api_paths'
 
 async function getChannelInform(channel: string) {
   try {
