+++ src/apis/post/getPostList.ts
@@ -1,5 +1,5 @@
 import axiosInstance from '../api'
-import { GET_POST_LIST_BY_CHANNEL_PATH } from '@/utils/api_paths'
+import { GET_POST_LIST_BY_CHANNEL_PATH } from '@/apis/api_paths'
 
 async function getPostList(channelId: string) {
   try {
