+++ src/apis/profile/updateName.ts
@@ -1,4 +1,4 @@
-import { UPDATE_MY_PROFILE_PATH } from '@/utils/api_paths'
+import { UPDATE_MY_PROFILE_PATH } from '@/apis/api_paths'
 import axiosInstance from '../api'
 
 interface Data {
