+++ src/apis/search/searchAll.ts
@@ -1,4 +1,4 @@
-import { SEARCH_ALL_PATH } from '@/utils/api_paths'
+import { SEARCH_ALL_PATH } from '@/apis/api_paths'
 import axiosInstance from '../api'
 
 interface SearchResponse {
