+++ src/apis/search/searchUser.ts
@@ -1,4 +1,4 @@
-import { SEARCH_USERS_PATH } from '@/utils/api_paths'
+import { SEARCH_USERS_PATH } from '@/apis/api_paths'
 import axiosInstance from '../api'
 
 interface SearchResponse {
