+++ src/apis/user/getAllUsers.ts
@@ -1,4 +1,4 @@
-import { GET_ALL_USER_LIST_PATH } from '@/utils/api_paths'
+import { GET_ALL_USER_LIST_PATH } from '@/apis/api_paths'
 import axiosInstance from '../api'
 
 interface UserResponse extends User {}
