+++ src/apis/user/getUserInform.ts
@@ -1,6 +1,6 @@
 import axiosInstance from '../api'
 
-import { GET_USER_PROFILE_PATH } from '@/utils/api_paths'
+import { GET_USER_PROFILE_PATH } from '@/apis/api_paths'
 
 async function getUserInform(id: string) {
   try {
